General enhancements.

+++ src/regular/modules/blocklist.py
@@ -12,37 +12,51 @@ async def is_user_admin(chat_id, user_id):
     return False
 
 async def sticker_block(m):
-    chat_id = str(m.chat.id).lstrip('-')
-    db = IMYDB(f'runtime/banned/{chat_id}_stickers.json')
-    banned = db.get('stickers', [])
+    try:
+        chat_id = str(m.chat.id).lstrip('-')
+        db = IMYDB(f'runtime/banned/{chat_id}_stickers.json')
+        banned = db.get('stickers', [])
+        self = await bot.get_me()
+        self = await bot.get_chat_member(m.chat.id, self.id)
 
-    if m.sticker and m.sticker.set_name in banned:
-        await bot.delete_message(m.chat.id, m.id)
+        if not self.can_delete_messages:
+            return
+
+        if m.sticker and m.sticker.set_name in banned:
+            await bot.delete_message(m.chat.id, m.id)
+
+    except Exception as error:
+        await log_error(bot, error, context_msg=m)
+        await bot.reply_to(m, "An error occurred.")
 
 async def block_set(m):
+    try:
 
-    if not await is_user_admin(m.chat.id, m.from_user.id):
-        await bot.reply_to(m, "Admin only.")
-        return
+        if not await is_user_admin(m.chat.id, m.from_user.id):
+            await bot.reply_to(m, "Admin only.")
+            return
 
-    if not m.reply_to_message:
-        await bot.reply_to(m, "Reply to a sticker.")
-        return
+        if not m.reply_to_message:
+            await bot.reply_to(m, "Reply to a sticker.")
+            return
 
-    elif not m.reply_to_message.sticker:
-        await bot.reply_to(m, "Reply to a sticker.")
-        return
+        elif not m.reply_to_message.sticker:
+            await bot.reply_to(m, "Reply to a sticker.")
+            return
 
-    set_name = m.reply_to_message.sticker.set_name
-    chat_id = str(m.chat.id).lstrip('-')
-    db = IMYDB(f'runtime/banned/{chat_id}_stickers.json')
-    banned = db.get('stickers', [])
+        set_name = m.reply_to_message.sticker.set_name
+        chat_id = str(m.chat.id).lstrip('-')
+        db = IMYDB(f'runtime/banned/{chat_id}_stickers.json')
+        banned = db.get('stickers', [])
 
-    if set_name not in banned:
-        banned.append(set_name)
-        db.set('stickers', banned)
+        if set_name not in banned:
+            banned.append(set_name)
+            db.set('stickers', banned)
 
-    await bot.reply_to(m, "Sticker set added to blacklist.")
+        await bot.reply_to(m, "Sticker set added to blacklist.")
+    except Exception as error:
+        await log_error(bot, error, context_msg=m)
+        await bot.reply_to(m, "An error occurred.")
 
 async def get_blacklist(m):
     chat_id = str(m.chat.id).lstrip('-')
@@ -53,8 +67,10 @@ async def get_blacklist(m):
         if stickers:
             output = '\n'.join([f"`{set_name}`" for set_name in stickers])
             await bot.send_message(m.chat.id, f"Banned sets:\n{output}", parse_mode="Markdown")
+
         else:
             await bot.send_message(m.chat.id, "No banned sets.")
+
     except Exception as error:
         await log_error(bot, error, context_msg=m)
         await bot.reply_to(m, "An error occurred.")
